% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_time.R
\name{adjust_time}
\alias{adjust_time}
\title{Adjust time}
\usage{
adjust_time(timestamp, timestamp_16, timezone = "UTC",
  origin = "1989-12-31  01:59:00", min_time = NULL, max_time = NULL)
}
\arguments{
\item{timestamp}{integer, the \code{timestamp[s]} field of FIT file}

\item{timestamp_16}{integer, the \code{timestamp_16[s]} field of FIT file}

\item{timezone}{character representing the timezone of input.}

\item{origin}{time origin.}

\item{min_time}{\code{[dttm]} minimum acceptable time, any output before this time
will be NA. If NULL (default) no minimum constrains will be applied.}

\item{max_time}{\code{[dttm]} maximum acceptable time, any output after this time
will be NA. If NULL (default) no maximum constrains will be applied .}
}
\value{
a \code{[dttm]} vector
}
\description{
Adjust time
}
