% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border_date.R
\name{border_date}
\alias{border_date}
\title{Create a border datetime}
\usage{
border_date(fit_date = NULL, side = c("low", "high"))
}
\arguments{
\item{fit_date}{\code{[dttm]} a dttm or NULL}

\item{side}{\code{[chr]}lower or upper}
}
\value{
a \code{[dttm]}
}
\description{
If a datetime is provided the function return that datetime, otherwise
it return a lower or upper border for datetime
}
