% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_monitoring.R
\name{import_monitoring}
\alias{import_monitoring}
\title{Import Monitoring FIT data}
\usage{
import_monitoring(base_path = ".", patient_name = NULL, fit_date = NULL,
  quietly = TRUE, verbose = FALSE)
}
\arguments{
\item{base_path}{the base path into start search the files}

\item{patient_name}{an optional patient's name (used as a subfolder of the
\code{base_path} provided)}

\item{fit_date}{an optional date used as a subfolder of the
\code{patient_name}
once (if provided) or directly as a subfolder of the \code{base_path}}

\item{quietly}{\code{[lgl]} Would you like to know what the function is doing?
(default is TRUE)}

\item{verbose}{\code{[lgl]} Would you like to know (also) what
\code{\link[readr]{read_csv}} is doing?}
}
\value{
a dataframe
}
\description{
the funciton aims to import the monitoring data from a FIT file. The intent
is to use it for a clinical project. Hence an optional structured path with
name and dates is provided as input.
}
\examples{
\dontrun{
 import_monitoring("data-raw", fit_date = "2018-06-02")
}
}
