% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/java_fit.R
\name{java_fit}
\alias{java_fit}
\title{FitCSVTools}
\usage{
java_fit()
}
\value{
\code{[chr]} path to the \code{FitCSVTool} program
}
\description{
This function possibly return the system-specific path to
the \code{FitCSVTool} program
}
