% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fit.R
\name{read_fit}
\alias{read_fit}
\title{Read FIT files}
\usage{
read_fit(path_to_fit, ..., tidy = TRUE, quietly = FALSE, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\description{
Read FIT files
}
\examples{
read_fit("./data-raw/2018-06-02/21188777597.fit")
}
