% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fit.R
\name{read_fit}
\alias{read_fit}
\title{Read FIT files}
\usage{
read_fit(path_to_fit, ..., quietly = FALSE, verbose = FALSE)
}
\arguments{
\item{path_to_fit}{\code{[chr]} the path to the FIT file}

\item{...}{additional argument passed to \code{\link[readr]{read_csv}}}

\item{quietly}{\code{[lgl]} Would you like to know what the function is doing?
(default is FALSE)}

\item{verbose}{\code{[lgl]} Would you like to know (also) what
\code{\link[readr]{read_csv}} is doing?}
}
\value{
a list of dataframe, each one for each kind of data stored into the
FIT file provided.
}
\description{
Read FIT files
}
\examples{
\dontrun{
  read_fit("./data-raw/2018-06-02/21188777597.fit")
}
}
